feat(proto): Use repeated fields
The use of repeated fields allows for supplying an array field, instead
of having to pass JSON everywhere. This will also make encoding and
decoding really fast on both client and server.

This change also renames the service to be shorter. Comments were
added for documentation when the Go client is generated.